remove unused type and move swapper types from types package to swapper package
