+++ packages/swapper/src/api.ts
@@ -1,15 +1,7 @@
 import { AssetId } from '@shapeshiftoss/caip'
 import { createErrorClass } from '@shapeshiftoss/errors'
 import { HDWallet } from '@shapeshiftoss/hdwallet-core'
-import {
-  ApprovalNeededOutput,
-  Asset,
-  ChainSpecific,
-  GetMinMaxInput,
-  MinMaxOutput,
-  SupportedChainIds,
-  SwapperType
-} from '@shapeshiftoss/types'
+import { Asset, ChainSpecific, SupportedChainIds, SwapperType } from '@shapeshiftoss/types'
 
 export const SwapError = createErrorClass('SwapError')
 
@@ -99,11 +91,6 @@ export type TradeResult = {
   txid: string
 }
 
-export type SwapSource = {
-  name: string
-  proportion: string
-}
-
 export type ApproveInfiniteInput<C extends SupportedChainIds> = {
   quote: TradeQuote<C>
   wallet: HDWallet
@@ -113,6 +100,28 @@ export type ApprovalNeededInput<C extends SupportedChainIds> = {
   quote: TradeQuote<C>
   wallet: HDWallet
 }
+
+export type SwapSource = {
+  name: string
+  proportion: string
+}
+
+export interface MinMaxOutput {
+  minimum: string
+  maximum: string
+}
+
+export type GetMinMaxInput = {
+  sellAsset: Asset
+  buyAsset: Asset
+}
+
+export type ApprovalNeededOutput = {
+  approvalNeeded: boolean
+  gas?: string
+  gasPrice?: string
+}
+
 
 // Swap Errors
 export enum SwapErrorTypes {
